Fix: readded BaseDeepNetwork
