+++ sktime/networks/base.py
@@ -1,6 +1,27 @@
 """Abstract base class for deep learning networks."""
-from abc import ABC
+
+__author__ = ["Withington", "TonyBagnall"]
+
+from abc import ABC, abstractmethod
+
+from sktime.base import BaseObject
 
 
-class BaseDeepNetworkPyTorch(ABC):
-    """Abstract base class for deep learning networks using pytorch."""
+class BaseDeepNetwork(BaseObject, ABC):
+    """Abstract base class for deep learning networks."""
+
+    @abstractmethod
+    def build_network(self, input_shape, **kwargs):
+        """Construct a network and return its input and output layers.
+
+        Parameters
+        ----------
+        input_shape : tuple
+            The shape of the data fed into the input layer
+
+        Returns
+        -------
+        input_layer : a keras layer
+        output_layer : a keras layer
+        """
+        ...
